ESQL: Fixup CASE tests
Move a bunch of the tests for CASE into their own file. They were in the
parameterized tests which we were running over and over and over again
but we really just want to run them once.

I'm doing this before I really go and fix #112359
